% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityflag.R
\name{flagR}
\alias{flagR}
\title{Quality flagger function}
\usage{
flagR(file)
}
\arguments{
\item{file}{A character string representing the full file path to raw csv data.}
}
\value{
A csv of the original data with quality issues flagged, a shape file
of the same and salinity, temperature and density profile plots.
}
\description{
\code{flagR} assesses the raw field sample data from a sonde and flags with a
numerical code any suspect samples in an output csv. OK data has a \code{QUAL}
code value of 10000. Quality issues are represented by codes that are
added to this value. Quality issues are:
\itemize{
\item Sample stability check (depth/Vpos +/- 0.2m variation) - \code{QUAL}
code 01000
\item Sample site check for out of sequence record - \code{QUAL} code 00100
\item Sample site check for duplicate record (within 0.1m) - \code{QUAL} code 00010
\item Transect check for expected and or incorrectly named samples
- \code{QUAL} cade 00001 }
}
\details{
The \code{flagR} function output csv has the original GPS co-ordinates
transformed from WGS84 to GDA94 MGA50. All eastings and northings are
therefore measured in metres. \code{flagR} also outputs an Esri shape file
with an attribute table created from the flagged raw data input. Lastly,
salinity, temperature and density profile plots, facetted by transect are
output in png format. All outputs are written to the working directory.
}
\examples{
\dontrun{
flagR("./my_raw_sonde_data.csv")
}
}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
