% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityflag.R
\name{imp_raw_csv}
\alias{imp_raw_csv}
\title{Import raw csv data from sonde}
\usage{
imp_raw_csv(file)
}
\arguments{
\item{file}{A character string of full file path to raw csv data.}
}
\value{
A tibble of the raw data with a \code{QUAL} variable set to 10000.
}
\description{
Import raw csv data from sonde
}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
