% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityflag.R
\name{old_data_updatR}
\alias{old_data_updatR}
\title{Update old sample site names to current versions}
\usage{
old_data_updatR(x)
}
\arguments{
\item{x}{A tibble output from running \code{\link{imp_raw_csv}}}
}
\value{
A tibble with site names in current format
}
\description{
Update old sample site names to current versions
}
