% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityflag.R
\name{plot_nameR}
\alias{plot_nameR}
\title{Helper function to reformat a file name.}
\usage{
plot_nameR(file)
}
\arguments{
\item{file}{A character string of full file path to raw csv data.}
}
\value{
A more human friendly plot title.
}
\description{
Helper function to reformat a file name.
}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
