% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityflag.R
\name{plotR}
\alias{plotR}
\title{Internal profile plot creator}
\usage{
plotR(x, file)
}
\arguments{
\item{x}{A tibble as exported from either \code{\link{imp_raw_csv}}
or \code{\link{old_data_updatR}}}

\item{file}{}
}
\value{
Three individual profile plots, facetted by transect, for salinity,
temperature and density (UNESCO formula).
}
\description{
Internal profile plot creator
}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
