% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityflag.R
\name{site_checkR}
\alias{site_checkR}
\title{Checks site samples for correct sequence and duplicate depth measurements.}
\usage{
site_checkR(x)
}
\arguments{
\item{x}{A tibble as exported from either \code{\link{imp_raw_csv}}
or \code{\link{old_data_updatR}}}
}
\value{
A tibble where \code{QUAL} variable has 100 added for an out of sequence
sample and/or 10 added for a duplicate.
}
\description{
Checks site samples for correct sequence and duplicate depth measurements.
}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
