% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityflag.R
\name{transect_checkR}
\alias{transect_checkR}
\title{Checks named samples are from expected transects and if named correctly.}
\usage{
transect_checkR(x)
}
\arguments{
\item{x}{A tibble as exported from either \code{\link{imp_raw_csv}}
or \code{\link{old_data_updatR}}}
}
\value{
A tibble where \code{QUAL} variable has 1 added for any sample not spatially
located in appropriate transect or sample has been named incorrectly.
}
\description{
Checks named samples are from expected transects and if named correctly.
}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
