% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityflag.R
\name{flagR}
\alias{flagR}
\title{Quality flagger function}
\usage{
flagR(file)
}
\arguments{
\item{file}{A character string of full file path to raw csv data.}
}
\value{
A csv of the original data with quality issues flagged, a shape file
of the same and salinity, temperature and density profile plots.
}
\description{
\code{flagR} assesses the raw field sample data from a sonde and flags with a
numerical code any suspect samples in an output csv. Suspect issues are:
\itemize{
\item Sample stability check (depth/Vpos +/-10\% variation) - \code{QUAL}
code 01000
\item Sample site check for out of sequence or duplicate (within 0.1m)
records - \code{QUAL} codes 00100 and 00010
\item Transect check for expected and or incorrectly named samples
- \code{QUAL} cade 00001 }
}
\details{
The \code{flagR} function output csv has the original GPS co-ordinates
transformed from WGS84 to GDA94 MGA50. All eastings and northings are
therefore measured in metres. \code{flagR} also outputs an Esri shape file
with an attribute table created from the flagged raw data input. Lastly,
salinity, temperature and density profile plots, facetted by transect are
output in png format. All outputs are written to the working directory.
}
\examples{
\dontrun{
flagR("./my_raw_sonde_data.csv")
}
}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
