% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityflag.R
\name{profile_plotR}
\alias{profile_plotR}
\title{Constructs profile plots for salinity, temperature and density.}
\usage{
profile_plotR(x, file)
}
\arguments{
\item{x}{A tibble as exported from either \code{\link{imp_raw_csv}}
or \code{\link{old_data_updatR}}}

\item{file}{A character string of full file path to raw csv data.}
}
\value{
Three individual profile plots, facetted by transect, for salinity,
temperature and density (UNESCO formula).
}
\description{
Constructs profile plots for salinity, temperature and density.
}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
