% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityflag.R
\name{profile_plotR}
\alias{profile_plotR}
\title{Constructs profile plots for salinity, temperature and density.}
\usage{
profile_plotR(file)
}
\arguments{
\item{file}{A character string. Full file path to raw csv data.}
}
\value{
Three individual profile plots, facetted by transect, for salinity,
temperature and density (UNESCO formula).
}
\description{
\code{profile_plotR} creates salinity, temperature and density profile plots.
It is used only on user edited quality flagged csv inputs and each
iteration will retain the input 'R' value in the plot file name.
}
\examples{
\dontrun{
profile_plotR("./my_raw_sonde_data.csv")
}
}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
